05_ecommerce_data_engineering_pipeline/ecomm_airflow/requirements.txt
# Airflow providers (compatible with Airflow 3.1 / Astro 3.1-7)
apache-airflow-providers-http
apache-airflow-providers-postgres
apache-airflow-providers-snowflake
apache-airflow-providers-google

# CloudSQL
cloud-sql-python-connector
sqlalchemy
pg8000

# Google Cloud
google-cloud-storage

# Cosmos (Airflow 3.x compatible)
astronomer-cosmos

# Snowflake core dependency (explicit but compatible)
snowflake-connector-python
snowflake-sqlalchemy

# Utilities (safe pins)
beautifulsoup4
requests
